Add test for absorbing retirement
